requirements.txt
langchain
langchain-ollama
langchain-groq
python-dotenv
streamlit
streamlit-chat
python-dotenv
